refactor(meetings): split ui_scan detector into focused modules
Lift the ~2.9k-line test block into tests.rs and slim the 668-line
run_meeting_detection_loop into a readable sequence of named steps, with
the meaty inline blocks extracted into focused sibling modules:

- pipeline.rs: discovering running meeting apps (DB browser hints + AX
  enumeration), scanning AX trees, and the scan-cadence / keep-alive policy
- events.rs: event-bus signal handlers (explicit stop, auto-end)
- lifecycle.rs: meeting start/end DB side effects, the no-apps path, shutdown
  end, and in-meeting flag sync
- tests.rs: the integration test suite

mod.rs drops from 3829 to 322 lines and the loop body reads as a sequence of
step functions. Pure reorganization — strictly behavior-preserving, no logic
changes. The cfg-gated discord_profile / find_running_meeting_apps /
resolve_browser_pid re-exports stay in mod.rs since other modules consume them.
